% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{loglik}
\alias{loglik}
\title{Calculates the analitical form of the log likelihood for Periodic Autocorrelated Logistic Regression}
\usage{
loglik(
  x,
  y,
  lower_bound = 1e-09,
  upper_bound = 1 - 1e-09,
  penalty_factor = 1,
  init_par,
  weights = NULL
)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response variable}

\item{lower_bound}{Probability lower boundary}

\item{upper_bound}{Probability upper boundary}

\item{penalty_factor}{Penalty factor for the individual coefficients}

\item{init_par}{Initial estimative of the parameters}

\item{weights}{Observation weights. Useful attribute for unbalanced data}
}
\value{

}
\description{
Calculates the analitical form of the log likelihood for Periodic Autocorrelated Logistic Regression
}
