% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{loglik_gradient}
\alias{loglik_gradient}
\title{Returns the gradient vector for log likelihood}
\usage{
loglik_gradient(x, y, ...)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response variable}

\item{...}{XXX}
}
\value{

}
\description{
Returns the gradient vector for log likelihood
}
