% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{loglik_hessian}
\alias{loglik_hessian}
\title{Returns the hessian matrix for log likelihood}
\usage{
loglik_hessian(x, y, ...)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response variable}

\item{...}{XXX}
}
\value{

}
\description{
Returns the hessian matrix for log likelihood
}
