% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rlap}
\alias{rlap}
\title{Fits the RLAP model}
\usage{
rlap(
  resp,
  covar,
  par0,
  lamb = 0,
  penalty_factor = 1,
  weights = NULL,
  trace = 0,
  lower_bound = 0.001,
  upper_bound = 0.999,
  maxiter = 500
)
}
\arguments{
\item{resp}{The response variable}

\item{covar}{The covariate matrix}

\item{par0}{Initial parameter}

\item{lamb}{some other parameter}

\item{penalty_factor}{also a parameter}

\item{weights}{Observation weights. Useful attribute for unbalanced data}

\item{trace}{XXX}

\item{lower_bound}{Probability lower boundary}

\item{upper_bound}{Probability upper boundary}

\item{maxiter}{Maximum number of iterations}
}
\value{

}
\description{
Fits the RLAP model
}
